make one component for backup warning
